Add helper methods for converting between 'slicec' and 'lsp_tower' types.
